"fix the error of numeric error in index.html, fix the error of data type of np in utils before prediction"

+++ utils.py
@@ -3,7 +3,7 @@ import numpy as np
 
 
 def preprocess(Open, High, Low, Volume):
-    test_data = np.array([[Open, High, Low, Volume]])
+    test_data = np.array([[Open, High, Low, Volume]], dtype=np.float64)
     trained_model = joblib.load('model.pkl')
     prediction = trained_model.predict(test_data)
